Move the change to SGXDCAMQuoteVerifierImpl
